+++ apps/src/tests/Test2466.tsx
@@ -1,8 +1,8 @@
-import { Header } from '@react-navigation/elements';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator, NativeStackNavigationProp } from '@react-navigation/native-stack';
-import React, { ForwardedRef } from 'react';
-import { findNodeHandle, GestureResponderEvent, Pressable, PressableProps, StyleSheet, Text, View } from 'react-native';
+import React from 'react';
+import { findNodeHandle, Text, View } from 'react-native';
+import PressableWithFeedback from '../shared/PressableWithFeedback';
 
 type StackParamList = {
   Home: undefined,
@@ -12,62 +12,7 @@ type RouteProps = {
   navigation: NativeStackNavigationProp<StackParamList>;
 }
 
-type PressableState = 'pressed-in' | 'pressed' | 'pressed-out'
-
-
 const Stack = createNativeStackNavigator<StackParamList>();
-
-const PressableWithFeedback = React.forwardRef((props: PressableProps, ref: ForwardedRef<View>): React.JSX.Element => {
-  const [pressedState, setPressedState] = React.useState<PressableState>('pressed-out');
-
-  const onPressInCallback = React.useCallback((e: GestureResponderEvent) => {
-    console.log('Pressable onPressIn', {
-      locationX: e.nativeEvent.locationX,
-      locationY: e.nativeEvent.locationY,
-      pageX: e.nativeEvent.pageX,
-      pageY: e.nativeEvent.pageY,
-    });
-    setPressedState('pressed-in');
-    props.onPressIn?.(e);
-  }, [props]);
-
-  const onPressCallback = React.useCallback((e: GestureResponderEvent) => {
-    console.log('Pressable onPress');
-    setPressedState('pressed');
-    props.onPress?.(e);
-  }, [props]);
-
-  const onPressOutCallback = React.useCallback((e: GestureResponderEvent) => {
-    console.log('Pressable onPressOut');
-    setPressedState('pressed-out');
-    props.onPressOut?.(e);
-  }, [props]);
-
-  const onResponderMoveCallback = React.useCallback((e: GestureResponderEvent) => {
-    console.log('Pressable onResponderMove');
-    props.onResponderMove?.(e);
-  }, [props]);
-
-  const contentsStyle = pressedState === 'pressed-out'
-    ? styles.pressablePressedOut
-    : (pressedState === 'pressed'
-      ? styles.pressablePressed
-      : styles.pressablePressedIn);
-
-  return (
-    <View ref={ref} style={[contentsStyle]}>
-      <Pressable
-        onPressIn={onPressInCallback}
-        onPress={onPressCallback}
-        onPressOut={onPressOutCallback}
-        onResponderMove={onResponderMoveCallback}
-      >
-        {props.children}
-      </Pressable>
-    </View>
-
-  );
-});
 
 function HeaderTitle(): React.JSX.Element {
   return (
@@ -138,18 +83,5 @@ function App(): React.JSX.Element {
     </NavigationContainer>
   );
 }
-
-const styles = StyleSheet.create({
-  pressablePressedIn: {
-    backgroundColor: 'lightsalmon',
-  },
-  pressablePressed: {
-    backgroundColor: 'crimson',
-  },
-  pressablePressedOut: {
-    backgroundColor: 'lightseagreen',
-  },
-});
-
 
 export default App;
